requirements.txt
streamlit~=1.44.0
pandas~=2.2.3
openpyxl~=3.1.5
python-docx~=1.1.2
docx2pdf~=0.1.8
PyPDF2~=3.0.1
pdf2docx~=0.5.8
PyMuPDF~=1.25.5
Pillow~=11.2.1
pydub~=0.25.1
audioop-lts~=0.2.1
google-generativeai~=0.8.3
python-dotenv~=1.0.1
pytest~=8.3.5
pytest-cov~=7.1.0
pypandoc~=1.13.0